backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
ultralytics==8.3.0
python-multipart==0.0.9
Pillow==10.4.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn==1.5.0
pandas==2.2.2
numpy>=1.24.0
openai==1.51.0
python-dotenv==1.0.1
